requirements.txt
elasticsearch
faiss-cpu
flask
flask-cors
pandas
python_dotenv
selenium
sentence-transformers
tensorflow
transformers
waitress
webdriver-manager